integrated todo CRUD front-to-back (auth jwt verified)

+++ repos/todos.js
@@ -7,19 +7,19 @@ const Todo = require('../models/todo');
 // http://localhost/api/todos/:userId
 router.get('/:userId', passport.authenticate('jwt', {session: false}), (req, res) => {
     const {userId} = req.params;
+
+    // if userId requested does not match token's user Id (ie. a user trying to get other user's todo list)
+    if (userId != req.user._id) {
+        res.json({success: false, msg: "Unauthorized"});
+        return;
+    }
     
     Todo.getAllTodosByUserId(userId, (err, todos) => {
-        if (err) {
-            res.json({
-                success: false,
-                msg: 'Failed to load todo items'
-            });
-        } else {
-            res.json({
-                success: true,
-                todos: todos
-            });
-        }
+        if (err) throw err;
+        res.json({
+            success: true,
+            todos: todos
+        });
     });
 });
 
@@ -27,6 +27,12 @@ router.get('/:userId', passport.authenticate('jwt', {session: false}), (req, res
 // http://localhost/api/todos
 router.post('/', passport.authenticate('jwt', {session: false}), (req, res) => {
     const { userId, title, description } = req.body;
+
+    // if userId requested does not match token's user Id (ie. a user trying to get other user's todo list)
+    if (userId != req.user._id) {
+        res.json({success: false, msg: "Unauthorized"});
+        return;
+    }
     
     const newTodo = new Todo({
         owner: userId,
@@ -48,15 +54,49 @@ router.post('/', passport.authenticate('jwt', {session: false}), (req, res) => {
 // http://localhost/api/todos/:todoId
 router.delete('/:todoId', passport.authenticate('jwt', { session: false }), (req, res) => {
     const { todoId } = req.params;
-    Todo.deleteTodo(todoId, (err, del) => {
+    const tokenUser = req.user;
+
+    Todo.getTodoByTodoId(todoId, (err, todo) => {
         if (err) throw err;
-        if (del.ok != 1) {
-            res.json({success: false, msg: "Unknown error occurred."});
-        } else if (del.deletedCount == 0) {
-            res.json({success: false, msg: "Todo with given ID does not exist."});
-        } else {
-            res.json({success: true, msg: "Successfully deleted a todo item."});
+        if (!todo.owner.equals(tokenUser._id)) {
+            res.json({success:false, msg: "Unauthorized"});
+            return;
         }
+        
+        Todo.deleteTodo(todoId, (err, del) => {
+            if (err) throw err;
+            if (del.ok != 1) {
+                res.json({success: false, msg: "Unknown error occurred."});
+            } else if (del.deletedCount == 0) {
+                res.json({success: false, msg: "Todo with given ID does not exist."});
+            } else {
+                res.json({success: true, msg: "Successfully deleted a todo item."});
+            }
+        });
+    });
+});
+
+// PUT
+// http://localhost/api/todos
+router.put('/', passport.authenticate('jwt', {session: false}), (req, res) => {
+    const { todoId, isDone } = req.body;
+    const tokenUser = req.user;
+
+    Todo.getTodoByTodoId(todoId, (err, todo) => {
+        if (todo === undefined) {
+            res.json({success: false, msg: "Todo with given ID not found"});
+            return;
+        }
+        if (err) throw err;
+        if (!todo.owner.equals(tokenUser._id)) {
+            res.json({success:false, msg: "Unauthorized"});
+            return;
+        }
+
+        Todo.updateTodo(todoId, isDone, (err, prev) => {
+            if (err) throw err;
+            res.json({success: true, prev: prev});
+        })
     });
 });
